+++ R/condense.R
@@ -8,16 +8,12 @@ NULL
 setMethod(
   f = "condense",
   signature = c("CompositionMatrix"),
-  definition = function(x, by = groups(x), drop = TRUE, ...) {
+  definition = function(x, by = groups(x), ...) {
     m <- nrow(x)
 
-    ## Validation
-    if (!is.list(by)) by <- list(by)
-    arkhe::assert_lengths(by, m)
-
     ## Grouping
-    index <- interaction(by, drop = drop, sep = "_")
-    if (length(unique(index)) == m) {
+    index <- as_groups(by)
+    if (nlevels(index) == 0 || nlevels(index) == m) {
       warning("Nothing to group by.", call. = FALSE)
       return(x)
     }
@@ -36,10 +32,11 @@ setMethod(
 
     tot <- tapply(X = totals(x), INDEX = index, FUN = mean, simplify = TRUE)
     grp <- groups(x)
+
     if (has_groups(grp)) grp <- flatten_chr(x = grp, by = index)
-    else grp <- rep(NA_character_, length(tot))
+    else grp <- rep(NA, length(tot))
 
     rownames(z) <- levels(index)
-    .CompositionMatrix(z, totals = as.numeric(tot), groups = grp)
+    .CompositionMatrix(z, totals = as.numeric(tot), groups = as_groups(grp))
   }
 )