+++ inst/tinytest/test_plot.R
@@ -18,16 +18,16 @@ if (at_home()) {
   expect_snapshot_plot(plot_hist, "plot_hist")
 
   # Barplot ====================================================================
-  plot_barplot <- function() barplot(coda, order_columns = FALSE)
+  plot_barplot <- function() barplot(coda, order_columns = FALSE, border = "black")
   expect_snapshot_plot(plot_barplot, "plot_barplot")
 
-  plot_barplot_order <- function() barplot(coda, order_columns = TRUE)
+  plot_barplot_order <- function() barplot(coda, order_columns = TRUE, border = "black")
   expect_snapshot_plot(plot_barplot_order, "plot_barplot_order_columns")
 
-  plot_barplot_order <- function() barplot(coda, order_rows = 2)
+  plot_barplot_order <- function() barplot(coda, order_rows = 2, border = "black")
   expect_snapshot_plot(plot_barplot_order, "plot_barplot_order_rows")
 
-  plot_barplot_group <- function() barplot(coda, by = rep(1:5, 5), order_columns = TRUE)
+  plot_barplot_group <- function() barplot(coda, by = rep(1:5, 5), order_columns = TRUE, border = "black")
   expect_snapshot_plot(plot_barplot_group, "plot_barplot_group")
 
   # Density ====================================================================